All buttons on end of level screen are functional. Still resolving all errors in the level loading process though.

+++ app/src/main/java/com/pandeagames/www/gutterballredux/Components/BodyComponent.java
@@ -83,6 +83,7 @@ public void createBody(Body body)
 		return body;
 	}
 	public void destroy(){
+		if(destroyed) {return;}
 		game.removeBodyComponent(this);
 		if(body!=null){
 			world.destroyBody(body);
